refactor(services): decompose god file into single-responsibility services
